[refactor] delete 패키지 생성하여 여행지 리뷰(PlaceReview) 삭제 DTO를 모아서 정리.
