--- src/main/java/com/haejwo/tripcometrue/domain/review/placereview/dto/response/DeletePlaceReviewResponseDto.java
@@ -1,4 +1,0 @@
-package com.haejwo.tripcometrue.domain.review.placereview.dto.response;
-
-public interface DeletePlaceReviewResponseDto {
-}
