+++ src/main/java/com/haejwo/tripcometrue/domain/review/placereview/dto/response/delete/DeletePlaceReviewResponseDto.java
@@ -1,0 +1,4 @@
+package com.haejwo.tripcometrue.domain.review.placereview.dto.response.delete;
+
+public interface DeletePlaceReviewResponseDto {
+}
